requirements.txt
mcp[cli]
python-dotenv
ipykernel
httpx
google-genai